ChangeTxPower use size_t instead of int

+++ changeTxPower.c
@@ -2,10 +2,10 @@
 #include <stdlib.h>
 #include <string.h>
 
-int isOnlyInts( int length, char *str )
+size_t isOnlyInts( size_t length, char *str )
 {
     /* must be ascii 48 to 57 inclusive */
-    int i;
+    size_t i;
     for( i=0; i<length; ++i )
     {
         if( str[i] == '\x00' )
@@ -17,7 +17,7 @@ int isOnlyInts( int length, char *str )
     return 1;
 }
 
-int main( int argc, char **argv )
+size_t main( size_t argc, char **argv )
 {
     if( argc == 1 )
     {
@@ -29,15 +29,15 @@ int main( int argc, char **argv )
         /* make sure argv[1] contains only integers for security.  Crap on the floor if it doesn't */
         if( isOnlyInts( strlen( argv[1] ), argv[1] ) )
         {
-            int strSize = strlen( argv[1] ) + 30;
+            size_t strSize = strlen( argv[1] ) + 30;
             char str[strSize];
-            int i;
+            size_t i;
             for( i=0; i<strSize; ++i )
                 str[i] = '\x00';
 
             snprintf( str, strSize - 1, "iwconfig wlan0 txpower %s", argv[1] );
 
-            int userid = getuid();
+            size_t userid = getuid();
             setuid( geteuid() );
             system( str );
             setuid( userid );
